Bring in applicable fixes from the main WavPack repository.
Note that these include security fixes for the demo command-line
programs and undefined behavior fixes for the library (left-shift
of negative numbers, etc). While bringing these in is recommended,
they are not strictly required for existing applications as there
have been no security vulnerabilities in the decoding library itself,
and any applications that use this library should be thoroughly
retested with these changes.